+++ log-writer/dfcti_log_writer.py
@@ -4,9 +4,11 @@
 import os
 import platform
 import numpy as np
-from numpy import random as rd
+from numpy.random import default_rng
 import time
 from datetime import datetime
+
+rd=default_rng()
 
 
 log_file_path = '/var/log/dfcti_system_logs.log'
@@ -140,4 +142,4 @@ class Write_Logs:
 
 
 MachineID.Generate_Machine_ID()
-Write_Logs.Write_Process(60, 1)
+Write_Logs.Write_Process(120, 1)