+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Latest updates 
 
+
+* Add commands in the GitHub Actions for running the log writer in the background
+* Changed the structure of the `resource` folder
+* Added class for attachment
+* Add alert class-module with support for mail alerts -> `dfcti_log_reader.py`
 * Add archive with the log writer script
 * Improves the pipeline for uploading artifact to the current release tag
 * Added correct versioning